feat: add note button
