+++ src/views/group/GroupViewSettingsTab.tsx
@@ -1,13 +1,15 @@
 import React from 'react'
 import { useGroupContext } from '../../contexts/group'
-import { useGetGroup } from '../../hooks/api/groups'
+import { useGetGroup, useUpdateGroup } from '../../hooks/api/groups'
 import GroupViewMenu from './GroupViewMenu'
 
 const GroupViewSettingsTab: React.FC = () => {
+  const [newGroupName, setNewGroupName] = React.useState('')
   const groupContext = useGroupContext()
   const getGroupQ = useGetGroup({ group_id: groupContext.groupID as string })
+  const updateGroupQ = useUpdateGroup()
   
-  return <div className='grid grid-rows-1 gap-8'>
+  return <div className='grid grid-rows-1 gap-4'>
     <GroupViewMenu activeTab={'settings'}>
       {
         getGroupQ.isSuccess ?
@@ -20,6 +22,13 @@ const GroupViewSettingsTab: React.FC = () => {
           <div className='skeleton w-56 h-7'></div>
       }
     </GroupViewMenu>
+    <div className='w-full flex flex-col'>
+      <label htmlFor='new-group-name-input'>Name</label>
+      <input id='new-group-name-input' type="text" value={newGroupName} onChange={(e) => setNewGroupName(e.target.value)} />
+      <button onClick={() => { updateGroupQ.mutate({ group: { id: groupContext.groupID as string, name: newGroupName }, update_mask: {paths: ['name'] }}) }}>
+          Submit
+      </button>
+    </div>
   </div>
 }
